BREAKING CHANGES: Please refer to the README as base commands have been
changed from `ConvertJSONtoTS` to `ConvertJSONtoLang {language}` and
vice versa for buffer files as well

+++ lua/json-to-types/module.lua
@@ -4,7 +4,7 @@ local utils = require("json-to-types.utils")
 ---@class CustomModule
 local M = {}
 
-M.write_types = function()
+M.write_types = function(target_language)
   local bufnr = vim.api.nvim_get_current_buf()
   local file_name = vim.api.nvim_buf_get_name(bufnr)
   local filetype = vim.api.nvim_get_option_value("filetype", { buf = bufnr })
@@ -17,18 +17,18 @@ M.write_types = function()
   if file then
     file:write(text)
     file:close()
-    helper.types_output(file_name)
+    helper.types_output(file_name, target_language)
     os.remove(file_path)
   else
     vim.notify("ERROR: Something went wrong")
   end
 end
 
-M.write_types_buffer = function()
+M.write_types_buffer = function(target_language)
   local bufnr = vim.api.nvim_get_current_buf()
   local file_name = vim.api.nvim_buf_get_name(bufnr)
-  local filetype = vim.api.nvim_get_option_value("filetype", { buf = bufnr })
-  local text = utils.buffer_to_string(filetype)
+  local filetype1 = vim.api.nvim_get_option_value("filetype", { buf = bufnr })
+  local text = utils.buffer_to_string(filetype1)
   if not text then
     return
   end
@@ -37,7 +37,10 @@ M.write_types_buffer = function()
   if file then
     file:write(text)
     file:close()
-    local types = helper.types_output_buffer(file_name)
+    local types = helper.types_output_buffer(file_name, target_language)
+    if not types then
+      return
+    end
     if string.find(types[1], Error_message) then
       vim.notify(Error_message)
       os.remove(file_path)
@@ -48,7 +51,7 @@ M.write_types_buffer = function()
         vim.api.nvim_command("botright vnew")
         buffer_number = vim.api.nvim_get_current_buf()
         vim.api.nvim_buf_set_name(buffer_number, types[2])
-        vim.api.nvim_set_option_value("filetype", "typescript", { buf = buffer_number })
+        vim.api.nvim_set_option_value("filetype", types[3], { buf = buffer_number })
       end
       local lines = vim.split(types[1], "\n")
       vim.api.nvim_buf_set_lines(buffer_number, 0, -1, false, lines)
